Override DoxygenDocumenter generate()
We don't want the default behaviour of generating into the sphinx
docstring directive, since this doesn't give us enough flexibility in
formatting (particularly for the <detaileddescription> output of
modules). We do need to tweak autodoxysummary (or autodoc?) to
generate our member documentation in between the <briefdescription>
and the <detaileddescription> of the module.