Better documentation of module methods
Instead of outputting methods at from autosummary, we find them with
autodoc for a given module. For each method, the documentation uses
the Fortran Sphinx domain, and creates field lists of parameters,
similar to Doxygen. The domain has no support for the "intent"
property of parameters, so this is only added as plain text.