Use smarter XML queries for speedups in autodoc
When Sphinx is started, we build one big tree from all the XML output
of Doxygen. This isn't a problem, except there were a number of
searches that started at the top of the tree, and carelessly didn't
restrict the scope of the search. Now we're trying to be a bit
smarter, such as searching from a parent element if we know the target
is a child, and not searching all the way to the bottom of the tree if
we know the target is near the top.

This gives a massive speedup in full documentation generation
time (~15-20 mins down to <5 mins).