Use <compoundname> for module names as link targets
To know how to generate the link for a target, we look for a name-like
field in the target node we've found by id. Methods use <name>, but
modules use <compoundname>, so we use the "kind" we've determined from
the link to tell us how to find the name of the target.