Use <description> to get return types, not <type>
Doxygen seems to be a little bit quirky when a function's return type
is e.g. "character(len=*)", and it thinks it has to refer to
"character" as a derived type from somewhere else in the source. This
gives a weird structure in the <type> field:

    <type><ref refid="..." kindref="member">character</ref>
          (len=*) function a_module::a_function</type>

I feel like parsing this correctly would be a bit of a
pain (flattening .text and .tail for all stuff inside <type>), when we
already have basically an exact duplicate of this information inside
the <description> tag, without any references, so let's just use that
instead!