Generate working Fortran links in autosummary table
Fortran links don't work with the full `module::function` qualified
name (probably due to how I'm setting up the directives, this can be
fixed!), but for the moment we'll just link to the working
version. Since we don't have the "any" role available, we're
hardcoding "func" for the moment, which isn't an ideal solution.