Generate summary for module types in autosummary
Luckily things are mostly self-conistent and this works without
requiring too much further modification: we get a table of types with
their description! The type pages themselves aren't being generated,
and the link to the type is still using the "func" role, so there are
some extra parentheses being appended.