Add a Documenter for types
This is modelled after the Documenter for methods, but it has to be
structured a little bit differently. Doxygen treats types as structs,
and puts them in their own XML file (and hence toplevel in the XML
tree), so we need to handle their structure and resolving the link
between their call from a module to their actual node.