Allow specification of member type in document_members()
To support the documentation of both types and methods within a module
page, we want to call document_members() multiple times within the
generate() method, for different types of members -- types and
functions, separately. We've restructured things to move the member
search from get_object_members() into document_members(), but this
probably isn't the right way to go.

This still needs some clean-up so we can filter members, etc. as was
being previously done.