Add calls to/from to function documentation
Within a function, we have access to `<references>` and
`<referencedby>` fields. These are also present in the Fortran domain
as `:callto:` and `:calledfrom:` fields.

There is a slight bug present here: references may refer to functions
or variables, but we are currently just assuming all references are to
functions. We could probably also generate a list of all variables
present in a module, then check the type of the reference (in
index.xml) and generate the correct link accordingly.