Generate files for standalone pages
We have a few files with `kind="page"`, which are standalone
documentation pages written in Doxygen. We can pretty easily process
them to ReST with our preexisting templating and XML processing
utilities. However, at this point we're starting to need a way to
split up the output content depending on type (so that we can still
use globs to generate tocs).