Support <programlisting> in documentation
There are a couple of types of preformatted text that can occur in
documentation, such as `<verbatim>`, or `<programlisting>`. We support
these in the general case by introducing the preformat_text() method,
which uses a bare ReST literal block. Program listings also specify
some extra nodes (for highlighting?), so we flatten() every line and
throw this information away.