Allow autosummary generation for different kinds of things
We want to allow autosummary generation for both modules and
pages (and perhaps other stuff, like types), but we don't want them to
all end up in the same generated/ directory. We revert to something of
the original behaviour of autosummary, by requiring separate .rst
files to control what gets generated, but we're still searching the
index file depending on the `kind` option.

At the moment, we have a hardcoded condition to not perform any
autosummary magic on pages, so we can just process them using a normal
toctree directive. On the plus side, we're back to having an
autosummary table generated for all the modules, instead of a long
toctree without the brief descriptions.